requirements.txt
gradio==3.26.0
opencv-python==4.7.0.72
numpy==1.24.0
matplotlib==3.7.1
scikit-learn==1.3.0
Pillow==9.4.0
